packages were changed, some test were added and some bugs were solved.

+++ src/main/java/org/example/controller/accounts/DeleteAccountServlet.java
@@ -1,9 +1,8 @@
-package org.example.servlet.accounts;
+package org.example.controller.accounts;
 
 
-import org.example.model.BankAccount;
-import org.example.service.BankAccountDAO;
-import org.example.service.UserDAO;
+import org.example.model.DAOs.BankAccountDAO;
+import org.example.service.DataValidator;
 import org.example.utils.ORMConfig;
 
 import javax.servlet.ServletException;
@@ -11,7 +10,6 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 @WebServlet(name = "accountDeleteServlet", value = "/accounts/delete")
@@ -25,15 +23,12 @@ public class DeleteAccountServlet extends HttpServlet {
     }
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        if (!DataValidator.hasRequiredParams(req,"id")){
+            resp.sendError(HttpServletResponse.SC_BAD_REQUEST, "Invalid form data");
+            return;
+        }
         int id = Integer.parseInt(req.getParameter("id"));
-//        HttpSession session = req.getSession();
-//        String userType = (String) session.getAttribute("userType");
-//        if (userType.endsWith("AccountHolder")){
             bankAccountDAO.deleteAccountById(id);
             resp.sendRedirect("showAllAccountsServlet");
-//        }
-//        else {
-//            resp.sendRedirect("/home?error=true");
-//        }
     }
 }
